requirements.txt
Django==2.1.2
django-leaflet==0.24.0
djangorestframework==3.8.2
djangorestframework-gis==0.13
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
